--- development_server/src/component_service.rs
@@ -1,1 +1,0 @@
-
